requirements.txt
fastapi
uvicorn
aiogram
httpx
google-generativeai
requests